refactor: move app_router to the app.api.routes module
